Refactor API routes to utilize state management for reactions, subscribers, and site stats; remove unused code and tests for fetch-popular-posts, fetch-post-analytics, related-posts, and subscribers; streamline reaction handling and exchange rate fetching logic.
