--- src/routes/api/analytics/server.test.ts
@@ -1,51 +1,0 @@
-import { describe, it } from 'vitest'
-
-describe('Analytics API', () => {
-	it.skip('should handle analytics data retrieval', () => {
-		// Test analytics data fetching
-		// Test response format
-		// Test data aggregation
-	})
-
-	it.skip('should validate date range parameters', () => {
-		// Test start/end date validation
-		// Test invalid date format handling
-		// Test future date rejection
-	})
-
-	it.skip('should handle authentication', () => {
-		// Test API key validation
-		// Test unauthorized access prevention
-		// Test admin-only access
-	})
-
-	it.skip('should implement caching', () => {
-		// Test cache headers
-		// Test cache invalidation
-		// Test cache hit/miss scenarios
-	})
-
-	it.skip('should handle external API integration', () => {
-		// Test third-party analytics API calls
-		// Test API rate limiting
-		// Test API error handling
-	})
-
-	it.skip('should aggregate data correctly', () => {
-		// Test data aggregation logic
-		// Test metric calculations
-		// Test data transformation
-	})
-
-	it.skip('should handle missing data', () => {
-		// Test empty dataset handling
-		// Test partial data scenarios
-		// Test default value provision
-	})
-
-	it.skip('should validate query parameters', () => {
-		// Test metric type validation
-		// Test filter parameter validation
-		// Test pagination parameters
-	})
-})
