--- src/routes/api/subscribers/server.test.ts
@@ -1,45 +1,0 @@
-import { describe, it } from 'vitest'
-
-describe('Subscribers API', () => {
-	it.skip('should handle GET requests for subscriber data', () => {
-		// Test subscriber list retrieval
-		// Test response format
-		// Test pagination if implemented
-	})
-
-	it.skip('should handle POST requests for new subscribers', () => {
-		// Test subscriber creation
-		// Test duplicate email handling
-		// Test validation
-	})
-
-	it.skip('should validate email addresses', () => {
-		// Test email format validation
-		// Test rejection of invalid emails
-		// Test error messages
-	})
-
-	it.skip('should handle authentication/authorization', () => {
-		// Test API key validation
-		// Test unauthorized access prevention
-		// Test admin-only operations
-	})
-
-	it.skip('should handle database errors', () => {
-		// Test database connection failures
-		// Test query execution errors
-		// Test transaction handling
-	})
-
-	it.skip('should implement rate limiting', () => {
-		// Test rate limit enforcement
-		// Test rate limit headers
-		// Test rate limit error responses
-	})
-
-	it.skip('should handle malformed requests', () => {
-		// Test invalid JSON handling
-		// Test missing required fields
-		// Test invalid HTTP methods
-	})
-})
