+++ src/routes/newsletter/+page.server.ts
@@ -1,10 +1,7 @@
-export const load = async ({ fetch }) => {
-	// Fetch newsletter subscriber count
-	const subscribers_promise = fetch(`api/subscribers`)
+import { get_subscriber_count } from '$lib/state/subscribers.svelte'
 
-	const subscribers_response = await subscribers_promise
-	const { newsletter_subscriber_count } =
-		await subscribers_response.json()
+export const load = async () => {
+	const { newsletter_subscriber_count } = await get_subscriber_count()
 	return {
 		newsletter_subscriber_count,
 	}
